Move overview stuff to its own package
