+++ src/main/java/ee/tuleva/onboarding/comparisons/overview/DummyAccountOverviewProvider.java
@@ -1,4 +1,4 @@
-package ee.tuleva.onboarding.comparisons;
+package ee.tuleva.onboarding.comparisons.overview;
 
 import ee.tuleva.onboarding.auth.principal.Person;
 import org.springframework.stereotype.Service;
