+++ src/main/java/ee/tuleva/onboarding/comparisons/overview/Transaction.java
@@ -1,4 +1,4 @@
-package ee.tuleva.onboarding.comparisons;
+package ee.tuleva.onboarding.comparisons.overview;
 
 import lombok.Value;
